# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "point api document. Error Codes: 80001 Invalid parameters; 80002 Internal server error; 80003 User already bound; 80004 Invite code already bound; 80005 Signature verification failed; 80006 Task verification failed; 80007 Invite code does not exist; 80008 Invite code type mismatch; 80009 Invite codes not enough",
        "title": "point API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/invite/bindDirect": {
            "post": {
                "description": "bind direct",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "bind direct",
                "parameters": [
                    {
                        "description": "bind direct req params",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReqBindDirect"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Rsp"
                        }
                    }
                }
            }
        },
        "/v1/invite/bindTask": {
            "post": {
                "description": "bind direct",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "bind direct",
                "parameters": [
                    {
                        "description": "bind direct req params",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReqBindDirect"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Rsp"
                        }
                    }
                }
            }
        },
        "/v1/invite/userStatus": {
            "get": {
                "description": "get user status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "get user status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Rsp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.RspUserStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ReqBindDirect": {
            "type": "object",
            "properties": {
                "invite_code": {
                    "type": "string"
                },
                "user_address": {
                    "type": "string"
                }
            }
        },
        "api.RspUserStatus": {
            "type": "object",
            "properties": {
                "bond_at": {
                    "type": "integer"
                },
                "bound": {
                    "type": "boolean"
                },
                "code_type": {
                    "type": "integer"
                },
                "invite_code": {
                    "type": "string"
                }
            }
        },
        "utils.Rsp": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}